Fix java.io.IOException: Error: merging entries into file states/18-01-15-21-00-06/MC_1.fp  java.io.IOException: Negative seek offset
[Bug][TLC]

+++ tlatools/src/tlc2/tool/fp/OffHeapDiskFPSet.java
@@ -1132,7 +1132,7 @@ public final class OffHeapDiskFPSet extends NonCheckpointableDiskFPSet implement
 
 	protected void writeIndex(long[] index, final RandomAccessFile raf, long length) throws IOException {
 		for (int i = 0; i < index.length; i++) {
-			long pos = Math.min(i * NumEntriesPerPage, length);
+			long pos = Math.min(((long) i) * NumEntriesPerPage, length);
 			raf.seek(pos * LongSize);
 			final long value = raf.readLong();
 			index[i] = value;
